# swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is the MCP Gateway API documentation.",
        "title": "MCP Gateway API",
        "contact": {
            "name": "Source Code",
            "url": "https://github.com/matthisholleville/mcp-gateway"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/v1/admin/claim-to-roles": {
            "get": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Get all claim to roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "claim to roles"
                ],
                "summary": "Get all claim to roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.ClaimToRolesConfig"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Upsert a claim to role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "claim to roles"
                ],
                "summary": "Upsert a claim to role",
                "parameters": [
                    {
                        "description": "Claim to role",
                        "name": "claimToRole",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.ClaimToRolesConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.ClaimToRolesConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/admin/claim-to-roles/{claimKey}/{claimValue}": {
            "delete": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Delete a claim to role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "claim to roles"
                ],
                "summary": "Delete a claim to role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Claim key",
                        "name": "claimKey",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Claim value",
                        "name": "claimValue",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/admin/proxies": {
            "get": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Get all proxies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Get all proxies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.ProxyConfig"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/admin/proxies/{name}": {
            "get": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Get a proxy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Get a proxy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proxy name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.ProxyConfig"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Upsert a proxy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Upsert a proxy",
                "parameters": [
                    {
                        "description": "Proxy",
                        "name": "proxy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.ProxyConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.ProxyConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Delete a proxy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "proxies"
                ],
                "summary": "Delete a proxy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Proxy name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/admin/roles": {
            "get": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Get all roles",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Get all roles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.RoleConfig"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Upsert a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Upsert a role",
                "parameters": [
                    {
                        "description": "Role",
                        "name": "role",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.RoleConfig"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.RoleConfig"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/admin/roles/{role}": {
            "delete": {
                "security": [
                    {
                        "Authentication": []
                    }
                ],
                "description": "Delete a role",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "roles"
                ],
                "summary": "Delete a role",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Role",
                        "name": "role",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "storage.ClaimToRolesConfig": {
            "type": "object",
            "properties": {
                "claim_key": {
                    "type": "string"
                },
                "claim_value": {
                    "type": "string"
                },
                "roles": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "storage.ConnectionConfig": {
            "type": "object",
            "properties": {
                "timeout": {
                    "$ref": "#/definitions/time.Duration"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "storage.ObjectType": {
            "type": "string",
            "enum": [
                "tools",
                "*"
            ],
            "x-enum-varnames": [
                "ObjectTypeTools",
                "ObjectTypeAll"
            ]
        },
        "storage.PermissionConfig": {
            "type": "object",
            "properties": {
                "object_name": {
                    "type": "string"
                },
                "object_type": {
                    "$ref": "#/definitions/storage.ObjectType"
                },
                "proxy": {
                    "type": "string"
                }
            }
        },
        "storage.ProxyAuthConfig": {
            "type": "object",
            "properties": {
                "header": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "storage.ProxyConfig": {
            "type": "object",
            "properties": {
                "auth": {
                    "$ref": "#/definitions/storage.ProxyAuthConfig"
                },
                "connection": {
                    "$ref": "#/definitions/storage.ConnectionConfig"
                },
                "name": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/storage.ProxyType"
                }
            }
        },
        "storage.ProxyType": {
            "type": "string",
            "enum": [
                "streamable-http"
            ],
            "x-enum-varnames": [
                "ProxyTypeStreamableHTTP"
            ]
        },
        "storage.RoleConfig": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "permissions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.PermissionConfig"
                    }
                }
            }
        },
        "time.Duration": {
            "type": "integer",
            "enum": [
                -9223372036854775808,
                9223372036854775807,
                1,
                1000,
                1000000,
                1000000000,
                60000000000,
                3600000000000
            ],
            "x-enum-varnames": [
                "minDuration",
                "maxDuration",
                "Nanosecond",
                "Microsecond",
                "Millisecond",
                "Second",
                "Minute",
                "Hour"
            ]
        }
    },
    "securityDefinitions": {
        "Authentication": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}